pyproject.toml
[project]
name = "populate-sql-server-to-mysql"
version = "0.1.0"
description = ""
authors = [
    {name = "SA"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv (>=1.1.1,<2.0.0)",
    "mysql-connector-python (>=9.4.0,<10.0.0)",
    "pyodbc (>=5.2.0,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
